implemented redux with multiple reducers
