+++ redux/actions/login-actions.ts
@@ -1,0 +1,5 @@
+export const login = () => ({
+    type: 'LOGIN',
+  });
+  export const logout = () => ({ type: 'LOGOUT' });
+
