+++ redux/types.ts
@@ -1,3 +1,9 @@
 export interface Theme {
     theme: String
 }
+export interface Action {
+    type:string
+}
+export interface Logged {
+    logged: boolean;
+}
